Store rrd step as long.

+++ opennms-rrd/opennms-rrd-cassandra/src/main/java/org/opennms/netmgt/rrd/cassandra/CassRrd.java
@@ -17,7 +17,7 @@ public class CassRrd {
 
     private String m_fileName;
 
-    private int m_step;
+    private long m_step;
 
     private List<String> m_dsNames;
 
@@ -72,8 +72,8 @@ public class CassRrd {
     public String getFileName() {
         return m_fileName;
     }
-
-    public int getStep() {
+    
+    public long getStep() {
         return m_step;
     }
